feat(app): Phase 3 verified - compiles and builds web
531 analyzer errors resolved to zero: DAO/model typing repairs, drift
conditional connection (native/web), glass theme+component API gaps,
webfeed replaced with direct package:xml parsing, collaboration remnants
removed from folders, sql.js loader for drift-on-web. flutter build web
succeeds; 19/19 widget tests pass.

+++ app/lib/ui/layouts/three_column_layout.dart
@@ -141,9 +141,9 @@ class _ThreeColumnLayoutState extends State<ThreeColumnLayout>
     final theme = GlassTheme.of(context);
     
     // Calculate actual widths
-    final actualLeftWidth = _leftCollapsed ? 60 : _leftWidth;
-    final actualRightWidth = _rightCollapsed ? 60 : _rightWidth;
-    final actualMiddleWidth = screenWidth - actualLeftWidth - actualRightWidth - (widget.dividerWidth * 2);
+    final actualLeftWidth = (_leftCollapsed ? 60 : _leftWidth).toDouble();
+    final actualRightWidth = (_rightCollapsed ? 60 : _rightWidth).toDouble();
+    final actualMiddleWidth = (screenWidth - actualLeftWidth - actualRightWidth - (widget.dividerWidth * 2)).toDouble();
     
     return Row(
       children: [
